Try to fix reconnection
If the connection to home assistant dies (e.g. if home assistant reboots),
we get into a state where sending doesn't work anymore and we keep spewing
error messages.  Try to change things so that the component handles the
message sending (to currently active servers), instead of having a single
server at a time.  Also move bluetooth scanning to start instead of on
connection so that we don't have to figure out what context it is in.